feat: improve AI card selection
Added difficulty levels affecting AI decisions.
AI now considers difficulty when selecting cards, using weighted probabilities for weaker moves.

+++ Source/include/GameEvents.h
@@ -43,7 +43,7 @@ struct EventCardHovered : GameEventBase {
 
 // Trzeba wybrac karte przy uzyciu Snatch
 struct EventSnatchTargetRequired : GameEventBase {
-	int opponentHandSize;
+	size_t opponentHandSize;
 };
 
 // Usunieta karta przez Snatch
